refactor(express-api): unified load-config approach
